:art: Added Books Class
